requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==2.1.7
djangorestframework==3.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Pygments==2.3.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1
